+++ src/compas_rhino/artists/volmeshartist.py
@@ -31,34 +31,90 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
     # ==========================================================================
 
     def clear(self):
+        """Delete all objects drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_vertices(self):
+        """Delete all vertices drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.vertex.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_edges(self):
+        """Delete all edges drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.edge.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_faces(self):
+        """Delete all faces drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.face.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_cells(self):
+        """Delete all cells drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.cell.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_vertexlabels(self):
+        """Delete all vertex labels drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.vertexlabel.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_edgelabels(self):
+        """Delete all edge labels drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.edgelabel.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
     def clear_facelabels(self):
+        """Delete all face labels drawn by this artist.
+
+        Returns
+        -------
+        None
+
+        """
         guids = compas_rhino.get_objects(name="{}.facelabel.*".format(self.volmesh.name))
         compas_rhino.delete_objects(guids, purge=True)
 
@@ -66,47 +122,26 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
     # draw
     # ==========================================================================
 
-    def draw(self, vertices=None, edges=None, faces=None, cells=None, vertexcolor=None, edgecolor=None, facecolor=None, cellcolor=None):
-        """Draw the network using the chosen visualisation settings.
+    def draw(self, cells=None, color=None):
+        """Draw a selection of cells.
 
         Parameters
         ----------
-        vertices : list[int], optional
-            A list of vertices to draw.
-            Default is None, in which case all vertices are drawn.
-        edges : list[tuple[int, int]], optional
-            A list of edges to draw.
-            The default is None, in which case all edges are drawn.
-        faces : list[int], optional
-            A selection of faces to draw.
-            The default is None, in which case all faces are drawn.
         cells : list[int], optional
-            A selection of cells to draw.
+            A list of cells to draw.
             The default is None, in which case all cells are drawn.
-        vertexcolor : :class:`~compas.colors.Color` | dict[int, :class:`~compas.colors.Color`], optional
-            The color of the vertices.
-            The default color is the value of :attr:`VolMeshArtist.default_vertexcolor`.
-        edgecolor : :class:`~compas.colors.Color` | dict[tuple[int, int], :class:`~compas.colors.Color`], optional
-            The color of the edges.
-            The default color is the value of :attr:`VolMeshArtist.default_edgecolor`.
-        facecolor : :class:`~compas.colors.Color` | dict[int, :class:`~compas.colors.Color`], optional
-            The color of the faces.
-            The default color is the value of :attr:`VolMeshArtist.default_facecolor`.
-        cellcolor : :class:`~compas.colors.Color` | dict[int, :class:`~compas.colors.Color`], optional
+        color : :class:`~compas.colors.Color` | dict[int, :class:`~compas.colors.Color`], optional
             The color of the cells.
-            The default color is the value of :attr:`VolMeshArtist.default_cellcolor`.
+            The default color is :attr:`VolMeshArtist.default_cellcolor`.
 
         Returns
         -------
         list[System.Guid]
             The GUIDs of the created Rhino objects.
+            Every cell is drawn as an individual mesh.
 
         """
-        guids = self.draw_vertices(vertices=vertices, color=vertexcolor)
-        guids += self.draw_edges(edges=edges, color=edgecolor)
-        guids += self.draw_faces(faces=faces, color=facecolor)
-        guids += self.draw_cells(cells=cells, color=cellcolor)
-        return guids
+        return self.draw_cells(cells=cells, color=color)
 
     def draw_vertices(self, vertices=None, color=None):
         """Draw a selection of vertices.
@@ -127,7 +162,7 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
 
         """
         self.vertex_color = color
-        vertices = vertices or list(self.volmesh.vertices())
+        vertices = vertices or self.vertices
         vertex_xyz = self.vertex_xyz
         points = []
         for vertex in vertices:
@@ -157,7 +192,7 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
 
         """
         self.edge_color = color
-        edges = edges or list(self.volmesh.edges())
+        edges = edges or self.edges
         vertex_xyz = self.vertex_xyz
         lines = []
         for edge in edges:
@@ -188,7 +223,7 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
 
         """
         self.face_color = color
-        faces = faces or list(self.volmesh.faces())
+        faces = faces or self.faces
         vertex_xyz = self.vertex_xyz
         facets = []
         for face in faces:
@@ -219,24 +254,23 @@ class VolMeshArtist(RhinoArtist, VolMeshArtist):
 
         """
         self.cell_color = color
-        cells = cells or list(self.volmesh.cells())
+        cells = cells or self.cells
         vertex_xyz = self.vertex_xyz
-        meshes = []
+        guids = []
         for cell in cells:
-            cell_faces = []
-            for fkey in self.volmesh.cell_faces(cell):
-                cell_faces.append({
-                    'points': [vertex_xyz[vertex] for vertex in self.volmesh.face_vertices(fkey)],
-                    'name': "{}.cell.{}.face.{}".format(self.volmesh.name, cell, fkey),
-                    'color': self.cell_color[cell].rgb255
-                })
-            guids = compas_rhino.draw_faces(cell_faces, layer=self.layer, clear=False, redraw=False)
-            guid = compas_rhino.rs.JoinMeshes(guids, delete_input=True)
-            compas_rhino.rs.ObjectLayer(guid, self.layer)
-            compas_rhino.rs.ObjectName(guid, '{}.cell.{}'.format(self.volmesh.name, cell))
-            compas_rhino.rs.ObjectColor(guid, self.cell_color[cell].rgb255)
-            meshes.append(guid)
-        return meshes
+            vertices = self.volmesh.cell_vertices(cell)
+            faces = self.volmesh.cell_faces(cell)
+            vertex_index = dict((vertex, index) for index, vertex in enumerate(vertices))
+            vertices = [vertex_xyz[vertex] for vertex in vertices]
+            faces = [[vertex_index[vertex] for vertex in self.halfface_vertices(face)] for face in faces]
+            guid = compas_rhino.draw_mesh(vertices,
+                                          faces,
+                                          layer=self.layer,
+                                          name="{}.cell.{}".format(self.volmesh.name, cell),
+                                          color=self.cell_color[cell].rgb255,
+                                          disjoint=True)
+            guids.append(guid)
+        return guids
 
     # ==========================================================================
     # draw labels